pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "jewishconversionacademy"
version = "0.1.0"
description = "Jewish Conversion Academy Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "sqlalchemy[asyncio]>=2.0.0,<3.0",
    "asyncpg>=0.30.0,<1.0",
    "alembic>=1.14.0,<2.0",
    "pydantic-settings>=2.6.0,<3.0",
    "firebase-admin>=6.6.0,<7.0",
    "google-cloud-storage>=2.18.0,<3.0",
    "google-cloud-tasks>=2.16.0,<3.0",
    "google-api-python-client>=2.150.0,<3.0",
    "python-multipart>=0.0.12,<1.0",
    "jinja2>=3.0.0,<4.0",
    "pydantic[email]>=2.0.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "pytest-mock>=3.14.0",
    "aiosqlite>=0.20.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
